+++ py_air_control_exporter/app.py
@@ -5,10 +5,12 @@ from werkzeug.middleware.dispatcher import DispatcherMiddleware
 from py_air_control_exporter import metrics
 
 
-def create_app():
+def create_app(host=None, protocol=None):
     app = Flask(__name__)
     metrics_collector_registry = prometheus_client.CollectorRegistry(auto_describe=True)
-    metrics_collector_registry.register(metrics.PyAirControlCollector())
+    metrics_collector_registry.register(
+        metrics.PyAirControlCollector(host=host, protocol=protocol)
+    )
     app.wsgi_app = DispatcherMiddleware(
         app.wsgi_app,
         {"/metrics": prometheus_client.make_wsgi_app(metrics_collector_registry)},