+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.1.5
 
+- Introduced the `py-air-control-exporter` command line tool.
+
 # 0.1.4
 
 - Packaging fixes: